src/models/deep_learning/rnn_classifier.py
"""
RNN classifier for encrypted traffic fingerprinting
"""

import numpy as np
from sklearn.base import BaseEstimator, ClassifierMixin

class RNNClassifier(BaseEstimator, ClassifierMixin):
    """RNN/LSTM classifier for sequence-based traffic classification"""
    
    def __init__(self, hidden_size=128, num_layers=2, num_classes=20):
        self.hidden_size = hidden_size
        self.num_layers = num_layers
        self.num_classes = num_classes
        self.is_fitted = False
        
    def fit(self, X, y):
        """Fit the RNN model - placeholder implementation"""
        self.classes_ = np.unique(y)
        self.is_fitted = True
        return self
    
    def predict(self, X):
        """Predict using the RNN model - placeholder implementation"""
        if not self.is_fitted:
            raise ValueError("Model must be fitted before prediction")
        
        return np.random.choice(self.classes_, size=len(X))
    
    def predict_proba(self, X):
        """Predict class probabilities - placeholder implementation"""
        if not self.is_fitted:
            raise ValueError("Model must be fitted before prediction")
        
        n_samples = len(X)
        n_classes = len(self.classes_)
        probs = np.random.dirichlet(np.ones(n_classes), size=n_samples)
        return probs
