requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Interactive plots
plotly>=5.0.0
dash>=2.0.0

# Network and packet processing
scapy>=2.4.5
pyshark>=0.4.6
dpkt>=1.9.7
ja3>=1.0.0

# Machine learning
xgboost>=1.5.0
lightgbm>=3.3.0
torch>=1.11.0
torchvision>=0.12.0
tensorflow>=2.10.0
keras>=2.10.0
imbalanced-learn>=0.9.0

# Web automation & scraping
selenium>=4.0.0
requests>=2.25.0

# Data processing utilities
pyyaml>=6.0
tqdm>=4.62.0
joblib>=1.1.0

# Utility / Dev
ipython>=7.30.0
jupyter>=1.0.0
colorama>=0.4.4
psutil>=5.8.0
pillow>=9.0.0

# GUI frameworks (CustomTkinter works, tkinter is built-in with Python)
customtkinter>=5.0.0

# Report generation
reportlab>=3.6.0
weasyprint>=56.0.0
jinja2>=3.1.0
